Move native query functions to new class
